+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased][unreleased]
+- Validate that statements do not contain fragment cycles
 
 ## [0.1.17] - 2016-10-14
 - Implement preliminary version of FieldsOnCorrectType validation
@@ -14,6 +15,7 @@ All notable changes to this project will be documented in this file. This change
 ## [0.1.16] - 2016-10-12
 ### BREAKING
 - [BREAKING] Change resolver to 3-parameter function
+
 ### CHANGED
 - Pass existing validation CATs (including error messages, mostly, code is still overly complex), including valid and invalid examples
 - Namespace schema (and statement variable) types for global spec naming